dbd_odbc: Fix SQL_INTEGER allocation size.
git-svn-id: https://svn.apache.org/repos/asf/apr/apr/trunk@1908501 13f79535-47bb-0310-9956-ffa450edef68

+++ dbd/apr_dbd_odbc.c
@@ -567,8 +567,8 @@ static SQLRETURN odbc_bind_param(apr_pool_t *pool,
                     (textmode ? atoi(args[*argp]) : *(short *)args[*argp]);
                 break;
             case SQL_INTEGER:
-                ptr = apr_palloc(pool, sizeof(int));
-                len = sizeof(int);
+                ptr = apr_palloc(pool, sizeof(long));
+                len = sizeof(long);
                 *(long *)ptr =
                     (textmode ? atol(args[*argp]) : *(long *)args[*argp]);
                 break;
